app/application/requirements.txt
Flask
azure-data-tables
pandas
azure-storage-blob
azure-identity
azure-mgmt-resource
azure-mgmt-storage